Break out and generalize
Inspired by parse_trans:plain_transform/2, the merlin:transform/3
accepts a transformer function and AST. Unlike the former, merlin
also accepts an additional state parameter that is passed along.

This makes it easy to carry some extra bits of information around,
like indentation level for a pretty printer.

Another difference is that merlin calls the tranformer both on the
way down, and on the way up, allowing the user to easily chose
between top-down or bottom-up, or a combination of both.

This is done thru an additional phase parameter, and it's strongly
recommended to include a _ catch all clause for future expansion.