Rewrite quote transform
This is a major rewrite and upgrade of the merlin_quote_transform.
It makes it possible to use top merl patterns not just as a top level
pattern in a function head, but arbitrarily nested in both function
heads and case clauses.

This makes it possible to carry a form around inside some sort of state
variable, like a record or map, match multiple merl patterns in the
same function head, and more.

The basic idea is to first divide the transform into two parts,
transforming function clauses to a case. Then transform these cases,
and any that the user might have written themselves, to something that
supports both Erlang and merl patterns.

The latter is done using a small helper function, just like how merl
does it. However it differs in the details. Instead of separate `fun`s
for the match and guard it uses only one per clause. For merl, that also
uses a template to do the matching, it makes more sense to have separate
`fun`s, but we can avoid it by controlling how the `fun`s return.

On success we return the result wrapped in `{ok, Value}`, and if the
clause didn't match `continue`. This is then used by
`merlin_quote_transform:switch/2` to know when to stop.

Finally, to make the stacktrace as good as possible, we raise
`function_clause` or `case_clause` as appropriately. Taking care to make
it look exactly like Erlangs builtin one.

-----------------------------------------------------------------------

It took a long time to get right, almost 3 weeks. But I think long term
it will be worth it. But it feels very good to put this behind me.