Simplify bindings management
First off, stick to only using `ordsets`, but still allow both atoms and
variable nodes.
Secondly, remove support for nesting using the `bindings` key.
The recursion makes the code harder to understand, and allow for weird
stuff like nesting a form inside an annotation.
Finally fix up all the other transforms that broke. The with transform
even got a lot simpler, so that's a win in my book.