Remove examples
They were useful for designing the interface, but now they have mostly
played their part. They might be brought back in some sort of `contrib`
directory, but that's for the future.

This is a breaking change, although I expect most users won't be
affected. On the flip side I think this makes it useable on OTP 24
again.