Small refactor of merlin_quote_transform
The `set_location' function isn't used, and if it turns out to be
needed, merlin_lib:set_annotation/3 already handles it.

The trick with using `switch_internal' has never been verified, and
besides those sort of things should be left to the runtime.

Finally, even though it is highly unlikely that someone overrides
`self/0', the parse transform should work regardless. We can easily
ensure that with a remote call to `erlang:self/0' instead.